requirements.txt
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
pdfplumber==0.11.0
python-docx==1.1.0
scikit-learn==1.4.1.post1
openai==1.55.3
plotly==5.19.0
